refactor(gps): consolidate TransReID backbone module
